% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{bbgr_mem_cache}
\alias{bbgr_mem_cache}
\title{Return the in-memory cache environment (advanced use)}
\usage{
bbgr_mem_cache()
}
\value{
An environment
}
\description{
Return the in-memory cache environment (advanced use)
}
