% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_cache.R
\name{bbgr_parquet_dir}
\alias{bbgr_parquet_dir}
\title{#' Cache directory used by bbgraphsR}
\usage{
bbgr_parquet_dir(path = NULL)
}
\value{
A length-1 character vector with the cache directory path.
}
\description{
Returns the path to the package cache directory (platform-specific),
typically under \code{rappdirs::user_cache_dir("bbgraphsR")}.
}
\seealso{
\code{\link[=bbgr_open_cache]{bbgr_open_cache()}}, \code{\link[=bbgr_parquet_clear]{bbgr_parquet_clear()}}
}
\keyword{internal}
