% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-country-flags.R
\docType{data}
\name{country_flags}
\alias{country_flags}
\title{Country Flag URLs for Use in MLB Debut Visualizations}
\format{
A data frame with 2 columns:
\describe{
\item{Country}{Character: country name as it appears in the player data}
\item{FlagURL}{Character: URL to a 40px PNG flag image}
}
}
\source{
\url{https://flagcdn.com}
}
\usage{
country_flags
}
\description{
A dataset mapping country names to 40px PNG flag images hosted on flagcdn.com.
Intended for use with functions like \code{viz_debut_running_total()}.
}
\keyword{datasets}
