% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_output.R
\name{.add_career_gcar}
\alias{.add_career_gcar}
\title{Add career-wide game counter (Gcar_real) across Regular+Post}
\usage{
.add_career_gcar(df)
}
\description{
Add career-wide game counter (Gcar_real) across Regular+Post
}
\keyword{(Gcar_real)}
\keyword{+}
\keyword{Add}
\keyword{Postseason}
\keyword{Regular}
\keyword{a}
\keyword{across}
\keyword{career-wide}
\keyword{game}
\keyword{index}
\keyword{internal}
