% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_cache.R
\docType{data}
\name{.bbgr_canonical_cols}
\alias{.bbgr_canonical_cols}
\title{Canonical (stable) column order used when returning data}
\format{
An object of class \code{character} of length 48.
}
\usage{
.bbgr_canonical_cols
}
\description{
Canonical (stable) column order used when returning data
}
\keyword{internal}
