% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-docs.R
\docType{data}
\name{teamcolors}
\alias{teamcolors}
\title{MLB team color palette}
\format{
A data frame with N rows and K columns.
}
\source{
Public color guides / team style guides.
}
\usage{
teamcolors
}
\description{
Primary/secondary colors for MLB teams.
}
\keyword{datasets}
