% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-docs.R
\docType{data}
\name{Teams}
\alias{Teams}
\title{MLB Teams table}
\format{
A data frame with N rows and K columns.
}
\source{
Compiled from Baseball-Reference / Lahman (see package README for details).
}
\usage{
Teams
}
\description{
Metadata about MLB teams (one row per team-season).
}
\keyword{datasets}
