% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-docs.R
\docType{data}
\name{TeamsFranchises}
\alias{TeamsFranchises}
\title{MLB Franchises table}
\format{
A data frame with N rows and K columns.
}
\source{
Compiled from Baseball-Reference / Lahman (see package README for details).
}
\usage{
TeamsFranchises
}
\description{
Metadata about MLB franchises (one row per franchise).
}
\keyword{datasets}
