% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_bbgraphs.R
\name{theme_bbgraphs}
\alias{theme_bbgraphs}
\title{My Brand ggplot2 Theme}
\usage{
theme_bbgraphs(font_size = 12, font_family = "Roboto")
}
\arguments{
\item{font_size}{Base font size.}

\item{font_family}{Base font family.}
}
\value{
A ggplot2 theme object.
}
\description{
A clean, consistent theme for my plots.
}
