% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_rd_team.R
\name{viz_rd_team}
\alias{viz_rd_team}
\title{Graph of Accumulated Runs Differential for a team over several seasons}
\usage{
viz_rd_team(team, start_year, end_year, highlight_year = NULL)
}
\arguments{
\item{team}{a string input, the Baseball Reference Team abbreviation}

\item{start_year}{a numeric value, the starting MLB season to be analyzed}

\item{end_year}{a numeric value, the ending MLB season to be analyzed}

\item{highlight_year}{a numeric value, one of the seasons (within the period start-end) which will be highlithed in the chart}
}
\value{
A areaspline & line types chart with the accumulated run differential for the Team along the seasons analyzed
}
\description{
Graph of Accumulated Runs Differential for a team over several seasons
}
\examples{
\dontrun{
viz_rd_team("TBR", 2019, 2023)
## returns an RD chart for TBR between 2015-2023, highlighting 2023
}
}
