% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_rd_wpct.R
\name{viz_rd_wpct}
\alias{viz_rd_wpct}
\title{Graph of Accumulated Runs Differential for teams}
\usage{
viz_rd_wpct(team, year)
}
\arguments{
\item{team}{a string input, Baseball Reference Team abbreviation, a division or the whole League}

\item{year}{a numeric value, MLB season to be analyzed}
}
\value{
A areaspline-type chart with the accumulated run differential for the Team(s) along the season analyzed
}
\description{
Function to plot the accumulated runs differential for an MLB team (or a group of them) on a specific MLB season
}
\examples{
viz_rd("AL West", 2021)
## returns an RD chart for all the AL West Teams in 2021, in descending order
\dontrun{
viz_rd_wpct("BOS", 2023)
## returns an RD chart for Boston Red Sox in the 2023 Season
viz_rd_wpct("NL Central", 2008)
## returns an RD chart for all the NL Central Teams in 2008, in descending order
 }
}
