% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_standings.R
\name{viz_standings}
\alias{viz_standings}
\title{Scrape MLB Standings on a specific season and visualize timelines of Games Behind (GB) and Winning percentage on any division or league}
\usage{
viz_standings(lg_div, year)
}
\arguments{
\item{lg_div}{a string input, Baseball Reference Team abbreviation or a division (e.g. NL East, NL Central, NL West, NL Overall, AL East, AL Central, AL West or AL Overall)}

\item{year}{a numeric value, MLB season to be analyzed}
}
\description{
This function allows you to scrape the standings from MLB teams for a specific season, and visualize two charts: Game Behind (GB) and Winning percentage of teams during that season.
}
\keyword{MLB,}
\keyword{standings}
