% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_wlp_season.R
\name{viz_wlp_season}
\alias{viz_wlp_season}
\title{Scrapes MLB Standings in one specific season and visualizes a timeline of the winning percentage for any division or a league}
\usage{
viz_wlp_season(lg_div, year, type = "ggplot")
}
\arguments{
\item{lg_div}{a string input, Baseball Reference Team abbreviation or a division (e.g. NL East, NL Central, NL West, NL Overall, AL East, AL Central, AL West, AL Overall or MLB)}

\item{year}{a numeric value, MLB season to be analyzed}

\item{type}{character, plotting backend to use: either
"ggplot" (default) for a static ggplot2 chart with team logos,
or "highcharter" for an interactive Highcharts visualization}
}
\description{
This function allows you to scrape the standings from MLB teams for a specific season, and visualize one chart of Winning percentage of teams during that season.
}
\keyword{MLB,}
\keyword{standings}
