% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_wlp_years.R
\name{viz_wlp_years}
\alias{viz_wlp_years}
\title{Scrapes MLB teams winning percentage from fangraphs.com along a period of years (at least 2 consecutive seasons) and visualizes a timeline for each team or franchise along that period}
\usage{
viz_wlp_years(start_season, end_season, fran_tm = "franchise")
}
\arguments{
\item{start_season}{a numeric value, first season you want to analyze (Should be at least a year before the last/current season)}

\item{end_season}{a numeric value, last season you want to analyze (Default value is 1 year after the input given in the 'start_season' argument)}

\item{fran_tm}{a string input, Option for indicating if the plot will be grouped by teams or by franchises (useful if period of season is longer)}
}
\description{
Scrapes MLB teams winning percentage from fangraphs.com along a period of years (at least 2 consecutive seasons) and visualizes a timeline for each team or franchise along that period
}
\examples{
\dontrun{
viz_wlp_years(2020, 2023)
}
}
\keyword{MLB,}
\keyword{performance}
