% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_wlp_years_gg.R
\name{viz_wlp_years_gg}
\alias{viz_wlp_years_gg}
\title{Visualize MLB Winning Percentage by Season (ggplot)}
\usage{
viz_wlp_years_gg(start_season, end_season, fran_tm = "franchise")
}
\arguments{
\item{start_season}{a numeric value, first season you want to analyze (Should be at least a year before the last/current season)}

\item{end_season}{a numeric value, last season you want to analyze (Default value is 1 year after the input given in the 'start_season' argument)}

\item{fran_tm}{a string input, Option for indicating if the plot will be grouped by teams or by franchises (useful if period of season is longer)}
}
\value{
A \code{ggplot} object created with \code{ggpubr::ggarrange()} that combines
one panel per requested team or franchise.
}
\description{
Produces the same winning-percentage timelines as \code{viz_wlp_years()} but
rendered with \code{ggplot2}, adding official team logos from \code{mlbplotR} in the
top-right corner of each panel.
}
