% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_scrape.R
\name{.keep_only_games}
\alias{.keep_only_games}
\title{Keep only actual game rows (drop summaries/totals)}
\usage{
.keep_only_games(df)
}
\description{
Keep only actual game rows (drop summaries/totals)
}
\keyword{#'}
\keyword{@noRd}
\keyword{internal}
