% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_debut_running_total.R
\name{viz_debut_running_total}
\alias{viz_debut_running_total}
\title{Plot Running Total of MLB Debuts by Country}
\usage{
viz_debut_running_total(countries)
}
\arguments{
\item{countries}{A character vector of country names (e.g. "Dominican Republic", "Venezuela")}
}
\value{
A ggplot2 object (line chart of cumulative MLB debuts).
}
\description{
Fetches player debut data from Baseball Reference and plots a cumulative line chart
of MLB player debuts over time. Countries can be compared via color-coded lines.
}
\examples{
\dontrun{
  plot_debut_running_total(c("Venezuela", "Dominican Republic", "Cuba"))
}

}
