% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_debut_years.R
\name{viz_debut_years}
\alias{viz_debut_years}
\title{Plot Yearly MLB Debuts by Country (Grouped Columns)}
\usage{
viz_debut_years(countries, start = NULL, end = NULL)
}
\arguments{
\item{countries}{A character vector of country names (e.g. "Dominican Republic", "Venezuela")}

\item{start}{Optional. An integer year (e.g. 1990) to begin the plot.}

\item{end}{Optional. An integer year (e.g. 2020) to end the plot.}
}
\value{
A ggplot2 object
}
\description{
Plots a grouped bar chart showing number of MLB player debuts per year, with a column per country.
}
