% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_rd.R
\name{viz_rd}
\alias{viz_rd}
\title{Graph of Accumulated Runs Differential}
\usage{
viz_rd(team, year)
}
\arguments{
\item{team}{a string input, Baseball Reference Team abbreviation, a division or the whole League}

\item{year}{a numeric value, MLB season to be analyzed}
}
\value{
A areaspline-type chart with the accumulated run differential for the Team(s) along the season analyzed

#' @examples
viz_rd("BOS", 2023)
viz-rd("AL West", 2021)
viz_rd("MLB", 2008)
}
\description{
Function to plot the accumulated runs differential for an MLB team (or a group of them) on a specific MLB season
}
