% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_wlp_years.R
\name{viz_wlp_years}
\alias{viz_wlp_years}
\title{Scrapes MLB teams winning percentage from fangraphs.com along a period of years (at least 2 consecutive seasons) and visualizes a timeline for each team or franchise along that period}
\usage{
viz_wlp_years(
  from_season,
  until_season = NULL,
  league = "all",
  fran_tm = "franchise"
)
}
\arguments{
\item{from_season}{a numeric value, first season you want to analyze (Default value is NULL, meaning one year before the 'until_season' argument)}

\item{until_season}{a numeric value, last season you want to analyze (in the worst case it must be the current/last MLB season)}

\item{league}{a string input, Option for limiting results to different leagues or overall results. Options are "al", "nl", or "all".}

\item{fran_tm}{a string input, Option for indicating if the plot will be grouped by teams or by franchises (useful if period of season is longer)}
}
\description{
Scrapes MLB teams winning percentage from fangraphs.com along a period of years (at least 2 consecutive seasons) and visualizes a timeline for each team or franchise along that period
}
\examples{
viz_wlp_years(2020, 2023)
}
\keyword{MLB,}
\keyword{performance}
